# Add a PROJ_DB_FAST_BUILD=ON/OFF CMake option (default OFF)