--- a/data/sql/analyze_vacuum.sql
+++ b/data/sql/analyze_vacuum.sql
@@ -1,0 +1,3 @@
+ANALYZE;
+
+VACUUM;