… data/CMakeLists.txt
… data/sql/commit.sql
… data/sql/conversion_triggers_hand_written.sql
… data/sql/versioned_auth_name_mapping.sql