@hobu 2024-10-16
Why would we add an option called `PROJ_DB_FAST_BUILD`? What is the reason for having `SLOW`?