@hobu 2024-10-16
My point was to ask why we should make this an option that users would have to make a decision about. I wonder if the behavior should be:

* default to `FAST` 
* add `PROJ_DB_EXTRA_VALIDATION=ON` for the `SLOW` mode