@rouault 2024-10-16
> That way we would have the best of both worlds.

just did that. Works just fine. Closing this PR as superseded per https://github.com/OSGeo/PROJ/pull/4280